refactor: extract api module
