--- colrev/packages/prospero/bin/prospero_results.json
@@ -1,6 +1,0 @@
-{
-    "results": [
-        {"id": 1, "title": "Protocol 1", "description": "Description of protocol 1"},
-        {"id": 2, "title": "Protocol 2", "description": "Description of protocol 2"}
-    ]
-}
